Save text box and toggle
Initial commit that brings in several base files.  The SetUp code has
been modified with a ton of comments and with two new UI elements.  The
first is a button that brings up a text box to change the base file
name.  The second is a toggle for the save button.