Moved to workign in seperate function
Moved to working ins eperate function to do the angle calibration so
that I can take advantage of Matlab recongizing
variables/functions/improper coding.  This will be copy pasted back into
the SetUp script once it is ready for testing.

Made untested progress to get and save a BMode and the RF data