Started looking into element calibration
Started looking into the element calibration, but found it was possible
to do it with a Signal Toolkit function.  Still playing around with the
function to see how it works and what logic I need to use to make it a
best fit and average.  Seems to do well with the individual signals now
though.