Hough lines adjustments
Started making changes to take advantage of the hough transform.  Right
now there's code to calculate and display the lines in a window around
the focus, but this is being tested through direct matlab statements and
not in the actual script.

If we want to use it in the script we need to do a few more steps to
determine the angle and send it further down the code.