Working!
The code now works properly, making a new BMode graph with an angle
overlay and tracking the angle properly.

There are some details, such as the axis on the graph or what angles can
be found, that can stand to be changed but it's functional enough for a
once every few month thing.